backend/requirements.txt
sqlalchemy
fastapi
PyPDF2
uvicorn
python-multipart
google-genai
python-dotenv
networkx